+++ app/Http/Controllers/EmployeeController.php
@@ -2,33 +2,103 @@
 
 namespace App\Http\Controllers;
 
-use Carbon\Carbon;
-use App\Models\User;
-use App\Models\Department;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Storage;
-use Yajra\DataTables\Facades\DataTables;
 use App\Http\Requests\StoreEmployeeRequest;
 use App\Http\Requests\UpdateEmployeeRequest;
-
+use App\Models\Department;
+use App\Models\User;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Storage;
+use Spatie\Permission\Models\Role;
+use Yajra\DataTables\Facades\DataTables;
 
 class EmployeeController extends Controller
 {
     public function index()
     {
+        if (!auth()->user()->can('view_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
         return view('employee.index');
+    }
+
+    public function ssd()
+    {
+        if(!auth()->user()->can('view_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
+        $employees = User::with('department');
+        return DataTables::of($employees)
+            ->filterColumn('department_name', function ($query, $keyword) {
+                $query->whereHas('department', function ($q1) use ($keyword) {
+                    $q1->where('name', 'like', '%' . $keyword . '%');
+                });
+            })
+            ->addColumn('plus-icon', function ($each) {
+                return null;
+            })
+            ->editColumn('profile_img', function ($each) {
+                return '<img src="' . $each->profile_img_path() . '" class="img-fluid img-thumbail profile_thumbnail rounded-circle" /> <p class="mt-2 text-muted">' . $each->name . '</p> ';
+            })
+            ->addColumn('department_name', function ($each) {
+                return $each->department->name ?? '-';
+            })
+            ->addColumn('role_name', function ($each) {
+                $output = "";
+                foreach ($each->roles as $role) {
+                    $output .= '<span class="badge badge-pill badge-success m-1">' . $role->name . '</span>';
+                }
+                return $output;
+            })
+            ->addColumn('action', function ($each) {
+                $edit_icon = '';
+                $detail_icon = '';
+                $delete_icon = '';
+                if(auth()->user()->can('edit_employee')) {
+                    $edit_icon = '<a href=" ' . route('employee.edit', $each->id) . ' " class="text-warning" title="edit">
+                    <i class="bx bxs-edit bx-sm"></i>
+                    </a>';
+                }
+
+                if(auth()->user()->can('view_employee')) {
+                    $detail_icon = '<a href=" ' . route('employee.show', $each->id) . ' " class="text-info mx-1" title="detail">
+                    <i class="bx bx-info-circle bx-sm"></i>
+                    </a>';
+                }
+
+                if(auth()->user()->can('delete_employee')) {
+                    $delete_icon = '<a href="' . route('employee.destroy', $each->id) . '" class="text-danger delete-btn" title="delete" data-id="' . $each->id . '">
+                    <i class="bx bxs-trash bx-sm"></i>
+                    </a>';
+                }
+
+                return '<div class="d-flex justify-content-center align-items-center align-middle">' . $edit_icon . $detail_icon . $delete_icon . '</div>';
+            })
+            ->editColumn('is_present', function ($each) {
+                return $each->is_present == 1 ? "<span class='badge badge-pill badge-success p-2'>Present</span>" : "<span class='badge badge-pill badge-danger p-2'>Leave</span>";
+            })
+            ->editColumn('updated_at', function ($each) {
+                return Carbon::parse($each->updated_at)->format("Y-m-d H:i:s");
+            })
+            ->rawColumns(['is_present', 'role_name', 'action', 'profile_img'])
+            ->make(true);
     }
 
     public function create()
     {
+        if(!auth()->user()->can('create_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
         $departments = Department::orderBy('name')->get();
-        return view('employee.create', compact('departments'));
+        $roles = Role::all();
+        return view('employee.create', compact('departments', 'roles'));
     }
 
     public function store(StoreEmployeeRequest $request)
     {
-
+        if(!auth()->user()->can('create_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
         if ($request->hasFile('profile_img')) {
             $profile_img_file = $request->file('profile_img');
             $profile_img_name = time() . '-' . uniqid() . '-' . $profile_img_file->getClientOriginalName();
@@ -53,25 +123,38 @@ class EmployeeController extends Controller
         $employee->is_present = $request->present;
 
         $employee->save();
+
+        $employee->syncRoles($request->roles);
+
         return redirect()->route('employee.index')->with('success', 'Employee is created successfully!');
     }
 
     public function show($id)
     {
+        if(!auth()->user()->can('view_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
         $employee = User::with('department')->findOrFail($id);
         return view('employee.show', compact('employee'));
     }
 
     public function edit($id)
     {
+        if(!auth()->user()->can('edit_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
         $employee = User::findOrFail($id);
         $departments = Department::orderBy('name')->get();
-
-        return view('employee.edit', compact('employee', 'departments'));
+        $roles = Role::all();
+        $old_roles = $employee->roles->pluck('id')->toArray();
+        return view('employee.edit', compact('employee', 'departments', 'roles', 'old_roles'));
     }
 
     public function update(UpdateEmployeeRequest $request, $id)
     {
+        if(!auth()->user()->can('edit_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
         $employee = User::findOrFail($id);
 
         $profile_img_name = $employee->profile_img;
@@ -98,57 +181,21 @@ class EmployeeController extends Controller
         $employee->is_present = $request->present;
 
         $employee->update();
+
+        $employee->syncRoles($request->roles);
+
         return redirect()->route('employee.index')->with('success', 'Employee is updated successfully!');
     }
 
     public function destroy($id)
     {
+        if(!auth()->user()->can('delete_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
         $user = User::findOrFail($id);
         $user->delete();
 
         return 'success';
     }
 
-
-    public function ssd()
-    {
-        $employees = User::with('department');
-        return DataTables::of($employees)
-            ->filterColumn('department_name', function($query, $keyword){
-                $query->whereHas('department', function($q1) use ($keyword) {
-                    $q1->where('name', 'like', '%'. $keyword .'%');
-                });
-            })
-            ->editColumn('profile_img', function($each) {
-                return '<img src="'. $each->profile_img_path() .'" class="img-fluid img-thumbail profile_thumbnail rounded-circle" /> <p class="mt-2 text-muted">'. $each->name .'</p> ';
-            })
-            ->addColumn('department_name', function ($each) {
-                return $each->department->name ?? '-';
-            })
-            ->addColumn('plus-icon', function ($each) {
-                return null;
-            })
-            ->addColumn('action', function ($each) {
-                $edit_icon = '<a href=" ' . route('employee.edit', $each->id) . ' " class="text-warning" title="edit">
-                    <i class="bx bxs-edit bx-sm"></i>
-                    </a>';
-                $detail_icon = '<a href=" ' . route('employee.show', $each->id) . ' " class="text-info mx-1" title="detail">
-                    <i class="bx bx-info-circle bx-sm"></i>
-                    </a>';
-
-                $delete_icon = '<a href="'. route('employee.destroy', $each->id) .'" class="text-danger delete-btn" title="delete" data-id="'.$each->id.'">
-                <i class="bx bxs-trash bx-sm"></i>
-                </a>';
-
-                return '<div class="d-flex justify-content-center align-items-center align-middle">' . $edit_icon . $detail_icon . $delete_icon . '</div>';
-            })
-            ->editColumn('is_present', function ($each) {
-                return $each->is_present == 1 ? "<span class='badge badge-pill badge-success p-2'>Present</span>" : "<span class='badge badge-pill badge-danger p-2'>Leave</span>";
-            })
-            ->editColumn('updated_at', function ($each) {
-                return Carbon::parse($each->updated_at)->format("Y-m-d H:i:s");
-            })
-            ->rawColumns(['is_present', 'action', 'profile_img'])
-            ->make(true);
-    }
 }
